--- src/main/java/Assistinator/AssitinatorExceptions.java
@@ -1,7 +1,0 @@
-package Assistinator;
-
-public class AssitinatorExceptions extends Exception{
-    public AssitinatorExceptions(String message) {
-        super(message);
-    }
-}
