+++ src/main/java/assistinator/AssitinatorExceptions.java
@@ -1,0 +1,10 @@
+package assistinator;
+
+/**
+ * Custom exception for Assistinator
+ */
+public class AssitinatorExceptions extends Exception {
+    public AssitinatorExceptions(String message) {
+        super(message);
+    }
+}
