+++ src/main/java/assistinator/Task.java
@@ -1,9 +1,16 @@
-package Assistinator;
+package assistinator;
 
+/**
+ * Represents a task
+ */
 public abstract class Task {
     protected String description;
     protected boolean isDone;
 
+    /**
+     * Initialise a task
+     * @param description Details of task
+     */
     public Task(String description) {
         this.description = description;
         this.isDone = false;
@@ -32,7 +39,7 @@ public abstract class Task {
      * @param keyword search keyword
      * @return whether description contains keyword
      */
-    public boolean contains(String keyword){
+    public boolean contains(String keyword) {
         return this.description.contains(keyword);
     }
 }
