+++ src/test/java/assistinator/AssistinatorTest.java
@@ -1,16 +1,16 @@
-package Assistinator;
-
-import org.junit.jupiter.api.Test;
+package assistinator;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 public class AssistinatorTest {
     @Test
-    public void dummyTest(){
+    public void dummyTest() {
         assertEquals(2, 2);
     }
 
     @Test
-    public void anotherDummyTest(){
+    public void anotherDummyTest() {
         assertEquals(4, 4);
     }
 }
