+++ src/test/java/assistinator/ParserTest.java
@@ -1,8 +1,9 @@
-package Assistinator;
+package assistinator;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 public class ParserTest {
     private Parser parser = new Parser();
